+++ lib/background/bg_service.dart
@@ -219,18 +219,43 @@ void onStart(ServiceInstance service) async {
 
   service.on("lock").listen((data) async {
     Logger("bgservice").info("Received lock command");
-    scooterService.lock();
+    if (scooterService.connected) {
+      setWidgetUnlocking(true);
+      await scooterService.lock(checkHandlebars: false);
+      setWidgetUnlocking(false);
+    } else {
+      // scan first, then lock
+      setWidgetScanning(true);
+      await attemptConnectionCycle();
+      setWidgetScanning(false);
+      if (scooterService.connected) {
+        setWidgetUnlocking(true);
+        await scooterService.lock(checkHandlebars: false);
+        Future.delayed(const Duration(seconds: 3), () {
+          setWidgetUnlocking(false);
+        });
+      }
+    }
   });
 
   service.on("unlock").listen((data) async {
     Logger("bgservice").info("Received unlock command");
     if (scooterService.connected) {
-      scooterService.unlock();
+      setWidgetUnlocking(true);
+      await scooterService.unlock(checkHandlebars: false);
+      setWidgetUnlocking(false);
     } else {
       // scan first, then unlock
       setWidgetScanning(true);
       await attemptConnectionCycle();
-      if (scooterService.connected) scooterService.unlock();
+      setWidgetScanning(false);
+      if (scooterService.connected) {
+        setWidgetUnlocking(true);
+        await scooterService.unlock(checkHandlebars: false);
+        Future.delayed(const Duration(seconds: 3), () {
+          setWidgetUnlocking(false);
+        });
+      }
     }
   });
 
@@ -242,6 +267,7 @@ void onStart(ServiceInstance service) async {
       // scan first, then open seat
       setWidgetScanning(true);
       await attemptConnectionCycle();
+      setWidgetScanning(false);
       if (scooterService.connected) scooterService.openSeat();
     }
   });