+++ changelog.md
@@ -1,4 +1,7 @@
 ### Changes:
-- Homescreen widget for Android
-- App can now connect to the Scooter in the background
+- Homescreen widget (Only on Android)
+- App can now connect to the Scooter in the background (Only on Android)
+- Advanced battery diagnostics
+- Fix for handlebar warning "Do not show again"
+- Major backend rework to improve performance
 - Various other minor fixes